adds joins

+++ Services/ReviewsService.cs
@@ -38,6 +38,10 @@ namespace AmaZen.Services
       {
         throw new Exception("Bad Request");
       }
+      if (original.CreatorId != update.CreatorId)
+      {
+        throw new Exception("Invalid Permissions");
+      }
       update.ProductId = original.ProductId;
       update.Title = update.Title != null ? update.Title : original.Title;
       update.Body = update.Body != null ? update.Body : original.Body;
@@ -45,12 +49,16 @@ namespace AmaZen.Services
       return _repo.Edit(update);
     }
 
-    internal String Delete(int id)
+    internal String Delete(int id, string userId)
     {
       var original = GetById(id);
       if (original == null)
       {
         throw new Exception("Bad Request");
+      }
+      if (original.CreatorId != userId)
+      {
+        throw new Exception("Invalid Permissions");
       }
       _repo.Delete(id);
       return "Successfully Deleted";
